webvtt/parsing/file-parsing/tests/stylesheets-newlines.html

<!doctype html>
<title>WebVTT parser test: stylesheets, rules</title>
<link rel="help" href="https://w3c.github.io/webvtt/#collect-a-webvtt-block">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var t = async_test('stylesheets, rules');
t.step(function(){
    var video = document.createElement('video');
    var track = document.createElement('track');
    assert_true('src' in track, 'track element not supported');
    track.src = 'support/stylesheets-newlines.vtt';
    track['default'] = true;
    track.kind = 'subtitles';
    track.onload = this.step_func(trackLoaded);
    track.onerror = this.step_func(trackError);
    video.appendChild(track);
    document.body.appendChild(video);
});

function trackLoaded(event) {
    var track = event.target;
    var video = track.parentNode;
    var cues = video.textTracks[0].cues;
    {
assert_equals(document.styleSheets.length, 1);

var styleSheet = document.styleSheets[0];
assert_equals(styleSheet.cssRules.length, 2);

var rule1 = styleSheet.cssRules[0];
assert_equals(rule1.type, CSSRule.STYLE_RULE);
assert_equals(rule1.selectorText, '::cue(#foo)');

var rule2 = styleSheet.cssRules[1];
assert_equals(rule2.type, CSSRule.STYLE_RULE);
assert_equals(rule2.selectorText, '.foo');

    }
    this.done();
}

function trackError(e) {
    assert_unreached('got unexpected error event: ' + JSON.stringify(e));
}
</script>
